add balance count to addTransaction
