+++ Backend/helpers/index.js
@@ -1,3 +1,1 @@
-export * from "./findTransactions.js";
-export * from "./getExpense.js";
-export * from "./getIncome.js";
+export * from "./transactionHelper.js";
